Move BaseListQuery out of ListQuery.ts
